+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Lazily-initiated static `fonts::FONT_HASHMAP` exported in top-level.
 - Lazily-initiated static `fonts::FONT_COLLECTION` exported in top-level.
 
+### Changed
+- Update to `font8x8 = 0.2` in Cargo.toml.
 ## [v0.1.11] - 2018-06-17
 ### Fixed
 - URLs to `examples/scroll-*.rs` on README.