+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ## [v0.2.0] - 2018-07-20
+### Fixed
+- Correct version in Usage section of README.md
+
+## [v0.2.0] - 2018-07-20
 ### Added
 - `lazy_static = "1.0"` to Cargo.toml.
 - Lazily-initiated static `fonts::FONT_HASHMAP` exported in top-level.